Refactor to use fast API.
