+++ Dockerfile
@@ -2,15 +2,11 @@ FROM python:3.6.12
 
 ENV SPACY_MODEL=en_core_web_sm
 
-RUN pip install flask flask-restful blackstone jsonpickle
+RUN pip install fastapi blackstone
 RUN pip install https://blackstone-model.s3-eu-west-1.amazonaws.com/en_blackstone_proto-0.0.1.tar.gz
 
-ADD server.py /
-ADD Legislation.py /
-ADD NamedEntity.py /
-ADD Abbreviation.py /
-ADD Sentence.py /
+EXPOSE 80
 
-EXPOSE 4449
+COPY ./app /app
 
-CMD [ "python", "./server.py" ]
+CMD ["python", "/app/main.py"]
