+++ app/models/NamedEntity.py
